Feature/layout improvements (#3)
* feat: Add Footer and Header components

The Footer and Header components are added to the app to provide a consistent layout and navigation elements.

* chore: Update tsconfig.json include and typeRoots

The tsconfig.json file is updated to include additional file patterns in the "include" section and to specify additional type roots in the "typeRoots" section. This ensures that the TypeScript compiler includes the necessary files and type definitions for the project.

* refactor: Rename Counter component file to follow naming conventions

The Counter component file is renamed from "app/counter.tsx" to "components/Counter.tsx" to adhere to the repository's naming conventions and improve code organization.

* feat: Export Header, Footer, and Counter components

This commit exports the Header, Footer, and Counter components from the components/index.tsx file. This allows other parts of the application to import and use these components easily.

* chore: Remove unused svg

* refactor: Update Home page layout and component structure

The Home page layout and component structure are refactored to improve code organization and readability. The Header, Footer, and Counter components are now imported from the components directory, and the main content is wrapped in a flex container. This change enhances the maintainability and scalability of the codebase.

* chore: Update npm dependencies

The package.json file is updated to include the latest versions of the following dependencies:
- @emotion/react: 11.11.4
- @emotion/styled: 11.11.5
- @mui/icons-material: 5.16.4
- @mui/material: 5.16.4

This ensures that the project is using the most up-to-date versions of these dependencies, which may include bug fixes, performance improvements, and new features.

* refactor: Update Home page layout and add ClickerCard component

The Home page layout is refactored to improve code organization and readability. The Header and Footer components are now imported from the components directory, and the main content is wrapped in a flex container. Additionally, a new ClickerCard component is added to the page, allowing users to interact with a clickable card. This change enhances the maintainability and scalability of the codebase and provides a better user experience.

* refactor: Add CurrencyProvider component for managing currency state

The CurrencyProvider component is added to the app to manage the state of the currency. This component provides a context that allows child components to access and modify the currency value. It also includes functionality to load the currency from localStorage, update it when it changes, and reset it to zero. This change improves the organization and maintainability of the codebase and enhances the user experience by providing a centralized way to manage the currency state.

* refactor: Remove Counter component

The Counter component is removed from the app as it is no longer needed. This change improves code organization and removes unnecessary code.

* refactor: Update Header component and add side menu functionality

The Header component is updated to use the Material-UI AppBar and Toolbar components for a more modern and responsive design. The side menu functionality is added, allowing users to toggle the menu open and close. This change improves the visual appearance and usability of the header.

* refactor: Update components/index.tsx to export ClickerCard instead of Counter

The components/index.tsx file is modified to export the ClickerCard component instead of the Counter component. This change improves code organization and ensures that the correct component is exported for usage in other parts of the application.